misc/HtmlTestSuite/results/test-0251.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Multi-column Layout Test: Default column-gap test animation</title>
<link href="mailto:rego@igalia.com" rel="author" title="Manuel Rego Casasnovas"/>
<link href="https://www.w3.org/TR/css-multicol-1/#column-gap" rel="help"/>
<link href="https://www.w3.org/TR/web-animations-1/#dom-animatable-animate" rel="help"/>
<meta content="Test checks that the default value for column-gap property, which is 'normal', is not interpolable." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
  @keyframes column-gap-to-100 {
    to { column-gap: 100px; }
  }

  #multicol {
    animation-name: column-gap-to-100;
    animation-duration: 2s;
    animation-delay: -1s;
    animation-play-state: paused;
  }
</style>
</head><body>
  <div id="multicol"/>
  <div id="log"/>

  <script>
    test(
      function(){
        var multicol = document.getElementById("multicol");
        assert_equals(getComputedStyle(multicol).columnGap, "100px");
      }, "Default column-gap is not interpolable");
  </script>
</body>
</html>
